refactor Ticket Templates Modal, Solve requested changes, and move files in diferent folders for templatesSettings, and PrinterSettings
