+++ client/src/components/pages/Store/Settings/TicketTemplate/TemplateElementRow.jsx
