+++ client/src/components/pages/Store/Settings/TicketTemplate/TemplateList.jsx
