+++ client/src/components/pages/Store/Settings/TicketTemplate/TemplatePreview.jsx
