+++ client/src/components/pages/Store/Settings/__tests__/PrinterAddForm.test.jsx
@@ -1,6 +1,6 @@
 import { fireEvent, render, screen } from "@testing-library/react";
 
-import { PrinterAddForm } from "../PrinterAddForm";
+import { PrinterAddForm } from "../PrinterSettings/PrinterAddForm";
 
 jest.mock("@heroui/react", () => ({
   Button: ({ onPress, isDisabled, children, ...props }) => (
